Move search hooks into hooks folder
