+++ src/common/hooks/useSearchType.js
